Delete files for languages at 0%
Just to save a bit of time on generating.
Anything with *something* translated I'll put the whole thing in to prevent 404s without needing to change languages, but if there's nothing translated there's no need for the files yet.
I'm keeping the languages though as I like having all of the languages shared between TWiLight and this.